refactor(corefoundation/base): more declarative approach to base abstract
